test: add grid_clear test case
