+++ HolubLife/test/GridTest.java
@@ -2,17 +2,15 @@ import com.holub.life.controller.Clock;
 import com.holub.life.controller.Universe;
 import com.holub.life.model.Neighborhood;
 import com.holub.life.model.Resident;
-
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 class GridTest {
-    @DisplayName("Oscillator single tick test")
+    @DisplayName("Clear Test")
     @Test
-    void OscillatorSingleTick() {
+    void Clear() {
         Clock clock = Clock.getInstance();
 
         Neighborhood oscillator = new Neighborhood(8, new Resident());
@@ -22,54 +20,20 @@ class GridTest {
         ((Resident) oscillator.getGrid()[1][3]).setAmALive(true);
 
         Universe oscUniverse = new Universe(clock, oscillator);
-        clock.tick();
+
+        oscUniverse.clear();
 
         assertAll(
-                // Before single tick
+                // After clear board
                 () -> assertFalse(((Resident) oscillator.getGrid()[1][1]).isAmAlive()),
-                () -> assertFalse(((Resident) oscillator.getGrid()[1][3]).isAmAlive()),
-
-                // After single tick
-                () -> assertTrue(((Resident) oscillator.getGrid()[0][2]).isAmAlive()),
-                () -> assertTrue(((Resident) oscillator.getGrid()[1][2]).isAmAlive()),
-                () -> assertTrue(((Resident) oscillator.getGrid()[2][2]).isAmAlive())
+                () -> assertFalse(((Resident) oscillator.getGrid()[1][2]).isAmAlive()),
+                () -> assertFalse(((Resident) oscillator.getGrid()[1][3]).isAmAlive())
         );
     }
 
-    @DisplayName("Oscillator double tick test")
-    @Test
-    void OscillatorDoubleTick() {
-        Clock clock = Clock.getInstance();
 
-        Neighborhood oscillator = new Neighborhood(8, new Resident());
-        oscillator.setAmActive(true);
-        ((Resident) oscillator.getGrid()[1][1]).setAmALive(true);
-        ((Resident) oscillator.getGrid()[1][2]).setAmALive(true);
-        ((Resident) oscillator.getGrid()[1][3]).setAmALive(true);
 
-        Universe oscUniverse = new Universe(clock, oscillator);
-        clock.tick();
-        clock.tick();
 
-        assertAll(
-                // After double tick
-                () -> assertTrue(((Resident) oscillator.getGrid()[1][1]).isAmAlive()),
-                () -> assertTrue(((Resident) oscillator.getGrid()[1][2]).isAmAlive()),
-                () -> assertTrue(((Resident) oscillator.getGrid()[1][3]).isAmAlive())
-        );
-    }
 
-    @DisplayName("Oscillator halt test")
-    @Test
-    void OscillatorHalt() {
-        Clock clock = Clock.getInstance();
 
-        Neighborhood oscillator = new Neighborhood(8, new Resident());
-        oscillator.setAmActive(true);
-        ((Resident) oscillator.getGrid()[1][1]).setAmALive(true);
-        ((Resident) oscillator.getGrid()[1][2]).setAmALive(true);
-        ((Resident) oscillator.getGrid()[1][3]).setAmALive(true);
-
-        Universe oscUniverse = new Universe(clock, oscillator);
-    }
 }
